fix(bit): re-apply the hoisted-resolution bridge after a layout-switching install
Gating the bridge on the recorded layout (607816bf5) opened a
first-transition gap: the install that switches a workspace onto the
global virtual store runs with bootstrap's gate still reflecting the
old layout, yet the same process goes on to reload envs and compile
from store slots without the bridge. The bridge now lives in
dependency-resolver as an idempotent module
(hoisted-resolution-bridge), deep-required lazily by bootstrap so CLI
startup never pays for the barrel, and the install flow re-applies it
right after the package manager finishes, the moment the target layout
is known.

The transition test also exposed the mirror gap: the running
installation itself may be store-linked (its packages living in
global-store slots) - then its own phantom requires need its private
hoist on NODE_PATH regardless of which workspace the process operates
on, and a workspace-only gate cut it off mid-transition. bootstrap now
bridges the installation first, judged by the same .modules.yaml
signal read at the installation root; a conventional project-local
installation reads project-local and stays unbridged, so nothing
changes for it or its children.

Verified end to end: a store-linked runner transitioned a workspace
local->global and back with resolution intact (the known env-config
retry aside), and steady-state installs stay green in both layouts.
